separate htt tutor object into file
